requirements.txt
pandas==1.1.3
streamlit==0.83.0
plotly==4.14.3
requests==2.24.0
python-dotenv==0.18.0
